platform/shared/requirements.txt
# Shared
pydantic>=2.0
pydantic-settings>=2.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
httpx>=0.27
python-multipart>=0.0.9
aiofiles>=24.1
redis>=5.0
boto3>=1.34
